Fix logging callbacks don't flush
If the logging callbacks don't flush, logging done by the child process doesn't appear for a long time, possibly when the buffer is full.

This causes any logging output of child processes to not appear until you exit the app for example, have had this discussion before, but I'm still convinced this is needed.

Flushing every log line of the child process is not a big deal because the child process hardly does any logging.